+++ src/main/java/pl/javastart/task/Point.java
@@ -1,19 +1,19 @@
 package pl.javastart.task;
 
 public class Point {
-    private double x;
-    private double y;
+    private int x;
+    private int y;
 
-    public Point(double x, double y) {
+    public Point(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 }
